requirements.txt
backports.strenum; python_version < '3.11'
